Add swagger upate and delete

+++ app/Http/Controllers/API/ProductController.php
@@ -1,5 +1,7 @@
 <?php
+
 namespace App\Http\Controllers\API;
+
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -29,7 +31,6 @@ class ProductController extends Controller
             "data" => $products
         ]);
     }
-
 
     /**
      * @OA\Post(
@@ -76,7 +77,7 @@ class ProductController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -86,7 +87,7 @@ class ProductController extends Controller
             'name' => 'required',
             'detail' => 'required'
         ]);
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
         $product = Product::create($input);
@@ -96,10 +97,25 @@ class ProductController extends Controller
             "data" => $product
         ]);
     }
+
+
+    /**
+     * @OA\Get(
+     *     tags={"Product"},
+     *     path="/api/products/{id}",
+     *      @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="Buscar por estado",
+     *         required=true,
+     *      ),
+     *     @OA\Response(response="200", description="List Products.")
+     * )
+     */
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -114,21 +130,74 @@ class ProductController extends Controller
             "data" => $product
         ]);
     }
+
+    /**
+     * @OA\Put (
+     ** path="/api/products/{id}",
+     *   tags={"Product"},
+     *
+     *  @OA\Parameter(
+     *      name="id",
+     *      in="path",
+     *      description="Buscar por estado",
+     *      required=true,
+     *   ),
+     *   @OA\Parameter(
+     *      name="name",
+     *      in="query",
+     *      required=true,
+     *      @OA\Schema(
+     *           type="string"
+     *      )
+     *   ),
+     *   @OA\Parameter(
+     *      name="detail",
+     *      in="query",
+     *      required=true,
+     *      @OA\Schema(
+     *          type="string"
+     *      )
+     *   ),
+     *   @OA\Response(
+     *      response=200,
+     *       description="Success",
+     *      @OA\MediaType(
+     *           mediaType="application/json",
+     *      )
+     *   ),
+     *   @OA\Response(
+     *      response=400,
+     *      description="Bad Request"
+     *   ),
+     *   @OA\Response(
+     *      response=404,
+     *      description="not found"
+     *   ),
+     *      @OA\Response(
+     *          response=403,
+     *          description="Forbidden"
+     *      )
+     *)
+     **/
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update($id, Request $request)
     {
+        $product = Product::find($id);
+        if (is_null($product)) {
+            return $this->sendError('Product not found.');
+        }
         $input = $request->all();
         $validator = Validator::make($input, [
             'name' => 'required',
             'detail' => 'required'
         ]);
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
         $product->name = $input['name'];
@@ -140,14 +209,51 @@ class ProductController extends Controller
             "data" => $product
         ]);
     }
+
+    /**
+     * @OA\Delete  (
+     ** path="/api/products/{id}",
+     *   tags={"Product"},
+     *
+     *  @OA\Parameter(
+     *      name="id",
+     *      in="path",
+     *      description="Buscar por estado",
+     *      required=true,
+     *   ),
+     *   @OA\Response(
+     *      response=200,
+     *       description="Success",
+     *      @OA\MediaType(
+     *           mediaType="application/json",
+     *      )
+     *   ),
+     *   @OA\Response(
+     *      response=400,
+     *      description="Bad Request"
+     *   ),
+     *   @OA\Response(
+     *      response=404,
+     *      description="not found"
+     *   ),
+     *      @OA\Response(
+     *          response=403,
+     *          description="Forbidden"
+     *      )
+     *)
+     **/
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Product $product)
+    public function destroy($id)
     {
+        $product = Product::find($id);
+        if (is_null($product)) {
+            return $this->sendError('Product not found.');
+        }
         $product->delete();
         return response()->json([
             "success" => true,
